make unique function to create dataset
